+++ CHANGELOG.md
@@ -6,6 +6,25 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 cluster API. See the changelog of the [cluster API](https://cluster-api.cyberfusion.nl/redoc#section/Changelog) for 
 detailed information.
 
+## [1.32.0]
+
+### Changed
+
+- Improve the ListFilter.
+- Add documentation about the filter.
+
+## [1.31.0]
+
+### Changed
+
+- Removed Psalm in favor of PHPStan.
+
+## [1.30.1]
+
+### Fixed
+
+- Fix the type of the unix name in the FPMPool model. Thanks to @szepeviktor.
+
 ## [1.30.0]
 
 ### Added